Rename python.rs to marker.rs
