+++ src/prelude.rs
@@ -16,9 +16,9 @@ pub use crate::conversion::{
 pub use crate::err::{PyErr, PyResult};
 pub use crate::gil::GILGuard;
 pub use crate::instance::{Py, PyObject};
+pub use crate::marker::Python;
 pub use crate::pycell::{PyCell, PyRef, PyRefMut};
 pub use crate::pyclass_init::PyClassInitializer;
-pub use crate::python::Python;
 pub use crate::types::{PyAny, PyModule};
 
 #[cfg(feature = "macros")]
